+++ Makefile
@@ -1,3 +1,14 @@
+.PHONY: alembic_revision
+alembic_revision:
+	docker-compose -f docker-compose.yml exec olive /bin/bash -c \
+	'alembic revision --autogenerate' && \
+	echo "Autogenerated Alembic revisions."
+
+.PHONY: alembic_upgrade
+alembic_upgrade:
+	docker-compose -f docker-compose.yml alembic upgrade head && \
+	echo "Upgraded database to the latest Alembic revision."
+
 .PHONY: build
 build:
 	docker-compose -f docker-compose.yml build && \
